lambdas/updateDownloadsData/requirements.txt
psycopg2-binary==2.8.6
boto3
moto